add HPSCREG_USER also to uwsgi service file
This is needed when adding the user
as part of an external url in the settings in Python.
The url is not used in our running uwsgi instance,
but the concatenation fails if the variable is not set.
We run the hPSCreg import
in a different service file
running directly on the docker image.

This needs a
sudo systemctl restart ims-uwsgi
on the ims and ims_staging hosts in the cloud,
if not done/detected by ansible automatically.
With
docker inspect ims-uwsgi | grep =
we can check if the environment is set up
correctly on ims and ims_staging.