fix ims image/container for proxy
We need to provide the proxy environment variables.
Additionally we need to provide the IP of proxy.charite.de

The IP of proxy.charite.de is not in public DNS.
The python docker image uses an ubuntu image as base,
which sets the google DNS server to 8.8.8.8
so proxy.charite.de is not known.

Since ansible docker_image can't use add-host,
we have to use the command module.