fix docker users
systemd is running a shell which doesn't allow commmand substitution.
So we run a shell just to set an environment variable with the user id.
This environment variable is set globally,
so it is also visible for other unit files.
I didn't want to put another bash call around ExecStart though.

Also note that the --user parameter of docker
expects a name known in the container (not on the host).
We want the id to have a name on the host though.
Giving an id works because these are shared between host and container.
So for that reason we want to translate
the user name to an id.