add another nginx layer for SSL/letsencrypt
We don't open a port for ims-nginx
(the nginx instance serving uwsgi),
but use another nginx instance reverse proxy.
This nginx instance was previously on another host (bastion).

With that instance we handle SSL/HTTPS certificates.
With this intermediate step it is also easier to run other services
on the same host later,
as the ims-nginx instance container is built directly
form sources in the ebisc repo.

Actual letsencrypt/certificate creation could not be tested yet,
as we still wait for a VM with a public DNS name.