can't use letsencrypt for charite (sub)domains
There is a CAA in DNS that only allows pki.dfn.de
to issue certificates for charite.de domains.

So we have to exclude these from letsencrypt.

As they are only used for testing internally,
it is fine to have them without a certificate.
(or rather not worth the hassle)