fix timeouts for nginx-proxy
proxy_read_timeout is only active
when set where proxy_pass is also set
(it seams).

Having it in the http section
keeps the default 60 second timeout!

The keepalive options are not active
in the server section
and only relevant in the upstream section
(which we don't use)

Getting the cell line ID CSV from the dashboard
takes more than 60 seconds (at least on local-dev)