common: enable to fully remove proxy
When an Ubuntu machine is installed with a proxy,
then /etc/apt/apt.conf is configured for the proxy.

We now also remove this when later changing use_proxy to false.

This should make it possible to remove the proxy from a machine
without creating a new VM.

This patch was taken from server-config repo as:
commit:e22ce46217d49a760218774cd06e43c56dcb2d40